refactor(runtime): extract runtime-event-history-search into a leaf
Move the history-search subsystem out of synthesis-cache.ts into
runtime-event-history-search.ts, keeping behavior neutral:

- extract searchRuntimeEventHistory, retrieveRuntimeEventHistoryAround,
  scoreRuntimeEventSearchHit, runtimeEventSearchText, and the
  RuntimeEventHistorySearchPolicy/Hit/AroundResult types to
  runtime-event-history-search.ts
- runtime-event-history-search depends only on the leaf helpers and
  tool-result-archive (isArchivedToolResultPlaceholder); synthesis-cache now
  imports runtimeEventSearchText forward from it (synthesis -> search ->
  archive), instead of keeping search private
- re-export all public symbols from context-budget.ts for import compatibility

synthesis-cache.ts drops to 752 lines and now contains only the synthesis
domain. No reverse imports.

Part of #1084.
